10_lesson/requirements.txt
selenium==4.15.0
pytest==7.4.3
allure-pytest==2.27.0
pytest-html==4.1.1
